server/pyproject.toml
[tool.poetry]
name = "video-embedding-db"
version = "0.1.0"
description = ""
authors = ["Chanwut (Mick) Kittivorawong <chanwutk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
clip = {git = "https://github.com/openai/CLIP.git", branch = "main"}
Flask = "^2.1.2"
Flask-Cors = "^3.0.10"
opencv-python = "^4.6.0"
numpy = "^1.23.0"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
